refactor(data): adapt IdentifyHandler to new VerbHandler abstracts
