+++ xoai-data-provider/src/test/java/io/gdcc/xoai/dataprovider/handlers/IdentifyHandlerTest.java
@@ -9,10 +9,10 @@
 package io.gdcc.xoai.dataprovider.handlers;
 
 import io.gdcc.xoai.dataprovider.exceptions.InternalOAIException;
-import io.gdcc.xoai.model.oaipmh.Identify;
+import io.gdcc.xoai.model.oaipmh.verbs.Identify;
 import org.junit.jupiter.api.Test;
 
-import static io.gdcc.xoai.model.oaipmh.Verb.Type.Identify;
+import static io.gdcc.xoai.model.oaipmh.verbs.Verb.Type.Identify;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 public class IdentifyHandlerTest extends AbstractHandlerTest {
     @Test
     public void validResponse() throws Exception {
-        Identify handle = new IdentifyHandler(aContext(), theRepository()).handle(a(request().withVerb(Identify)));
+        Identify handle = new IdentifyHandler(aContext(), theRepository()).handle(request().withVerb(Identify));
         String result = write(handle);
 
         assertThat(result, xPath("//repositoryName", is(equalTo(theRepositoryConfiguration().getRepositoryName()))));
